chore: reorganizar estrutura de importação e mover utilitários para diretórios institucionais
