--- app/api/data/preset/route.ts
@@ -1,18 +1,0 @@
-import { NextResponse } from "next/server";
-import path from "path";
-import fs from "fs";
-
-export async function GET() {
-  try {
-    const filePath = path.join(process.cwd(), "public", "data", "preset.json");
-    const jsonData = fs.readFileSync(filePath, "utf-8");
-    const data = JSON.parse(jsonData);
-    return NextResponse.json(data);
-  } catch (err) {
-    console.error("Erro ao ler JSON:", err);
-    return NextResponse.json(
-      { error: "Não foi possível carregar o JSON" },
-      { status: 500 }
-    );
-  }
-}
