+++ components/FormMission/RangeInput.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 export function RangeInput({
   missionId,
   index,
