--- config/folder-icons.ts
@@ -1,60 +1,0 @@
-import {
-  Archive,
-  Atom,
-  BookOpen,
-  Brain,
-  Briefcase,
-  Bug,
-  ClipboardList,
-  Code2,
-  FileText,
-  Flag,
-  Folder,
-  Gamepad2,
-  GraduationCap,
-  HeartPulse,
-  Image,
-  LayoutDashboard,
-  Lightbulb,
-  Microscope,
-  Music2,
-  Rocket,
-  Settings2,
-  ShieldCheck,
-  Star,
-  TestTube2,
-  Trophy,
-  Wrench,
-  type LucideIcon,
-} from "lucide-react";
-
-export const FOLDER_ICONS = {
-  folder: Folder,
-  code: Code2,
-  tests: TestTube2,
-  trophy: Trophy,
-  docs: FileText,
-  archive: Archive,
-  idea: Lightbulb,
-  project: Rocket,
-  settings: Settings2,
-  bug: Bug,
-  science: Microscope,
-  robot: Brain,
-  engineering: Wrench,
-  dashboard: LayoutDashboard,
-  study: GraduationCap,
-  image: Image,
-  music: Music2,
-  business: Briefcase,
-  health: HeartPulse,
-  security: ShieldCheck,
-  clipboard: ClipboardList,
-  game: Gamepad2,
-  atom: Atom,
-  strategy: Flag,
-  notes: BookOpen,
-  favorite: Star,
-} satisfies Record<string, LucideIcon>;
-
-export type FolderIconName = keyof typeof FOLDER_ICONS;
