--- public/data/preset.json
@@ -1,38 +1,0 @@
-[
-  {
-    "id_theme": 1,
-    "name": "Masterpiece Tema",
-    "background_url": "/images/showLive/backgrounds/background_masterpiece.png",
-    "colors": ["#ff5733", "#33ff57", "#ffffff"]
-  },
-  {
-    "id_theme": 2,
-    "name": "Submerged Tema",
-    "background_url": "/images/showLive/backgrounds/background_submerged.png",
-    "colors": ["#1abc9c", "#16a085", "#2ecc71"]
-  },
-  {
-    "id_theme": 3,
-    "name": "Unearthed Tema",
-    "background_url": "/images/showLive/backgrounds/background_unearthed.png",
-    "colors": ["#e74c3c", "#c0392b", "#ffffff"]
-  },
-  {
-    "id_theme": 4,
-    "name": "Dark Neon",
-    "background_url": "/images/presets/dark_neon.jpg",
-    "colors": ["#ff007c", "#00e5ff", "#ffffff"]
-  },
-  {
-    "id_theme": 5,
-    "name": "Natureza",
-    "background_url": "/images/presets/nature.webp",
-    "colors": ["#2ecc71", "#27ae60", "#145a32"]
-  },
-  {
-    "id_theme": 6,
-    "name": "Pôr do Sol",
-    "background_url": "/images/presets/sunset.webp",
-    "colors": ["#ff9a76", "#ff6b6b", "#c44536"]
-  }
-]
