--- types/Features.ts
@@ -1,7 +1,0 @@
-import type React from 'react';
-
-export interface Feature {
-  title: string;
-  text: string;
-  icon: React.ReactNode;
-}
