--- utils/formatters.ts
@@ -1,8 +1,0 @@
-export function formatDateTimeISO(iso?: string) {
-  if (!iso) return "";
-  const d = new Date(iso);
-  return `${d.toLocaleDateString("pt-BR")} ${d.toLocaleTimeString("pt-BR", {
-    hour: "2-digit",
-    minute: "2-digit",
-  })}`;
-}
