--- utils/helpers.ts
@@ -1,14 +1,0 @@
-export function getGreetingPrefix(): string {
-  const h = new Date().getHours()
-  if (h < 12) return "Bom dia"
-  if (h < 18) return "Boa tarde"
-  return "Boa noite"
-}
-
-export function formatDate(): string {
-  return new Date().toLocaleDateString("pt-BR", {
-    weekday: "long",
-    day: "numeric",
-    month: "long",
-  })
-}
